+++ sdk/servicebus/azure-servicebus/azure/servicebus/_servicebus_session_receiver.py
@@ -154,6 +154,9 @@ class ServiceBusSessionReceiver(ServiceBusReceiver, SessionReceiverMixin):
          within its request to the service.
         :rtype: ~azure.servicebus.ServiceBusSessionReceiver
 
+        :raises ~azure.servicebus.ServiceBusAuthenticationError: Indicates an issue in token/identity validity.
+        :raises ~azure.servicebus.ServiceBusAuthorizationError: Indicates an access/rights related failure.
+
         .. admonition:: Example:
 
             .. literalinclude:: ../samples/sync_samples/sample_code_servicebus.py