+++ sdk/storage/azure-storage-queue/azure/storage/queue/_shared/authentication.py
@@ -79,7 +79,9 @@ class SharedKeyCredentialPolicy(SansIOHTTPPolicy):
         uri_path = urlparse(request.http_request.url).path
         try:
             if isinstance(request.context.transport, AioHttpTransport) or \
-                    isinstance(getattr(request.context.transport, "_transport", None), AioHttpTransport):
+                    isinstance(getattr(request.context.transport, "_transport", None), AioHttpTransport) or \
+                    isinstance(getattr(getattr(request.context.transport, "_transport", None), "_transport", None),
+                               AioHttpTransport):
                 uri_path = URL(uri_path)
                 return '/' + self.account_name + str(uri_path)
         except TypeError: