+++ sdk/tables/azure-data-tables/azure/data/tables/aio/_table_client_async.py
@@ -20,17 +20,16 @@ from azure.core.exceptions import ResourceNotFoundError, HttpResponseError
 from azure.core.tracing.decorator import distributed_trace
 from azure.core.tracing.decorator_async import distributed_trace_async
 
-from .. import VERSION
 from .._base_client import parse_connection_str
 from .._entity import TableEntity
 from .._generated.aio import AzureTable
 from .._generated.models import SignedIdentifier, TableProperties, QueryOptions
-from .._models import AccessPolicy, TableItem
+from .._models import AccessPolicy
 from .._serialize import serialize_iso
 from .._deserialize import _return_headers_and_deserialized
 from .._error import _process_table_error
 from .._models import UpdateMode
-from .._deserialize import _convert_to_entity
+from .._deserialize import _convert_to_entity, _trim_service_metadata
 from .._serialize import _add_entity_properties, _get_match_headers
 from .._table_client_base import TableClientBase
 from ._base_client_async import AsyncStorageAccountHostsMixin
@@ -71,7 +70,6 @@ class TableClient(AsyncStorageAccountHostsMixin, TableClientBase):
             account_url, table_name=table_name, credential=credential, loop=loop, **kwargs
         )
         self._client = AzureTable(self.url, pipeline=self._pipeline, loop=loop)
-        self._client._config.version = kwargs.get('api_version', VERSION)  # pylint: disable = W0212
         self._loop = loop
 
     @classmethod
@@ -193,16 +191,18 @@ class TableClient(AsyncStorageAccountHostsMixin, TableClientBase):
             self,
             **kwargs  # type: Any
     ):
-        # type: (...) -> TableItem
+        # type: (...) -> Dict[str,str]
         """Creates a new table under the given account.
-        :return: Table created
-        :rtype: TableItem
+        :return: Dictionary of operation metadata returned from service
+        :rtype: dict[str,str]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         table_properties = TableProperties(table_name=self.table_name, **kwargs)
         try:
-            table = await self._client.table.create(table_properties)
-            return TableItem(table)
+            metadata, _ = await self._client.table.create(
+                table_properties,
+                cls=kwargs.pop('cls', _return_headers_and_deserialized))
+            return _trim_service_metadata(metadata)
         except HttpResponseError as error:
             _process_table_error(error)
 
@@ -240,7 +240,7 @@ class TableClient(AsyncStorageAccountHostsMixin, TableClientBase):
         :rtype: None
         :raises: ~azure.core.exceptions.HttpResponseError
         """
-        if_match, if_not_match = _get_match_headers(kwargs=dict(kwargs, etag=kwargs.pop('etag', None),
+        if_match, _ = _get_match_headers(kwargs=dict(kwargs, etag=kwargs.pop('etag', None),
                                                                 match_condition=kwargs.pop('match_condition', None)),
                                                     etag_param='etag', match_param='match_condition')
         try:
@@ -248,7 +248,7 @@ class TableClient(AsyncStorageAccountHostsMixin, TableClientBase):
                 table=self.table_name,
                 partition_key=partition_key,
                 row_key=row_key,
-                if_match=if_match or if_not_match or '*',
+                if_match=if_match or '*',
                 **kwargs)
         except HttpResponseError as error:
             _process_table_error(error)
@@ -256,42 +256,40 @@ class TableClient(AsyncStorageAccountHostsMixin, TableClientBase):
     @distributed_trace_async
     async def create_entity(
             self,
-            entity,  # type: Union[TableEntity, dict[str,str]]
+            entity,  # type: Union[TableEntity, Dict[str,str]]
             **kwargs  # type: Any
     ):
-        # type: (...) -> TableEntity
+        # type: (...) -> Dict[str,str]
         """Insert entity in a table.
         :param entity: The properties for the table entity.
         :type entity: dict[str, str]
-        :return: TableEntity mapping str to azure.data.tables.EntityProperty
-        :rtype: ~azure.data.tables.TableEntity
+        :return: Dictionary of operation metadata returned from service
+        :rtype: dict[str,str]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
-
-        if entity:
-            if "PartitionKey" in entity and "RowKey" in entity:
-                entity = _add_entity_properties(entity)
-            else:
-                raise ValueError('PartitionKey and RowKey were not provided in entity')
+        if "PartitionKey" in entity and "RowKey" in entity:
+            entity = _add_entity_properties(entity)
+        else:
+            raise ValueError('PartitionKey and RowKey were not provided in entity')
         try:
-            inserted_entity = await self._client.table.insert_entity(
+            metadata, _ = await self._client.table.insert_entity(
                 table=self.table_name,
                 table_entity_properties=entity,
+                cls=kwargs.pop('cls', _return_headers_and_deserialized),
                 **kwargs
             )
-            properties = _convert_to_entity(inserted_entity)
-            return properties
+            return _trim_service_metadata(metadata)
         except ResourceNotFoundError as error:
             _process_table_error(error)
 
     @distributed_trace_async
     async def update_entity(
             self,
-            entity,  # type: Union[TableEntity, dict[str,str]]
+            entity,  # type: Union[TableEntity, Dict[str,str]]
             mode=UpdateMode.MERGE,  # type: UpdateMode
             **kwargs  # type: Any
     ):
-        # type: (...) -> None
+        # type: (...) -> Dict[str,str]
         """Update entity in a table.
         :param mode: Merge or Replace entity
         :type mode: ~azure.data.tables.UpdateMode
@@ -305,11 +303,11 @@ class TableClient(AsyncStorageAccountHostsMixin, TableClientBase):
         :type etag: str
         :param match_condition: MatchCondition
         :type match_condition: ~azure.core.MatchConditions
-        :return: None
-        :rtype: None
+        :return: Dictionary of operation metadata returned from service
+        :rtype: dict[str,str]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
-        if_match, if_not_match = _get_match_headers(kwargs=dict(kwargs, etag=kwargs.pop('etag', None),
+        if_match, _ = _get_match_headers(kwargs=dict(kwargs, etag=kwargs.pop('etag', None),
                                                                 match_condition=kwargs.pop('match_condition', None)),
                                                     etag_param='etag', match_param='match_condition')
 
@@ -317,20 +315,27 @@ class TableClient(AsyncStorageAccountHostsMixin, TableClientBase):
         row_key = entity['RowKey']
         entity = _add_entity_properties(entity)
         try:
+            metadata = None
             if mode is UpdateMode.REPLACE:
-                await self._client.table.update_entity(
+                metadata, _ = await self._client.table.update_entity(
                     table=self.table_name,
                     partition_key=partition_key,
                     row_key=row_key,
                     table_entity_properties=entity,
-                    if_match=if_match or if_not_match or "*",
+                    if_match=if_match or "*",
+                    cls=kwargs.pop('cls', _return_headers_and_deserialized),
                     **kwargs)
             elif mode is UpdateMode.MERGE:
-                await self._client.table.merge_entity(table=self.table_name, partition_key=partition_key,
-                                                      row_key=row_key, if_match=if_match or if_not_match or "*",
-                                                      table_entity_properties=entity, **kwargs)
+                metadata, _ = await self._client.table.merge_entity(
+                    table=self.table_name,
+                    partition_key=partition_key,
+                    row_key=row_key,
+                    if_match=if_match or "*",
+                    cls=kwargs.pop('cls', _return_headers_and_deserialized),
+                    table_entity_properties=entity, **kwargs)
             else:
                 raise ValueError('Mode type is not supported')
+            return _trim_service_metadata(metadata)
         except HttpResponseError as error:
             _process_table_error(error)
 
@@ -429,19 +434,19 @@ class TableClient(AsyncStorageAccountHostsMixin, TableClientBase):
     @distributed_trace_async
     async def upsert_entity(
             self,
-            entity,  # type: Union[TableEntity, dict[str,str]]
+            entity,  # type: Union[TableEntity, Dict[str,str]]
             mode=UpdateMode.MERGE,  # type: UpdateMode
             **kwargs  # type: Any
     ):
-        # type: (...) -> None
+        # type: (...) -> Dict[str,str]
 
         """Update/Merge or Insert entity into table.
         :param mode: Merge or Replace and Insert on fail
         :type mode: ~azure.data.tables.UpdateMode
         :param entity: The properties for the table entity.
         :type entity: dict[str, str]
-        :return: Entity mapping str to azure.data.tables.EntityProperty or None
-        :rtype: None
+        :return: Dictionary of operation metadata returned from service
+        :rtype: dict[str,str]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
 
@@ -450,25 +455,30 @@ class TableClient(AsyncStorageAccountHostsMixin, TableClientBase):
         entity = _add_entity_properties(entity)
 
         try:
+            metadata = None
             if mode is UpdateMode.MERGE:
-                await self._client.table.merge_entity(
+                metadata, _ = await self._client.table.merge_entity(
                     table=self.table_name,
                     partition_key=partition_key,
                     row_key=row_key,
                     table_entity_properties=entity,
+                    cls=kwargs.pop('cls', _return_headers_and_deserialized),
                     **kwargs
                 )
             elif mode is UpdateMode.REPLACE:
-                await self._client.table.update_entity(
+                metadata, _ = await self._client.table.update_entity(
                     table=self.table_name,
                     partition_key=partition_key,
                     row_key=row_key,
                     table_entity_properties=entity,
+                    cls=kwargs.pop('cls', _return_headers_and_deserialized),
                     **kwargs)
             else:
-                raise ValueError('Mode type is not supported')
+                raise ValueError("""Update mode {} is not supported.
+                    For a list of supported modes see the UpdateMode enum""".format(mode))
+            return _trim_service_metadata(metadata)
         except ResourceNotFoundError:
-            await self.create_entity(
+            return await self.create_entity(
                 partition_key=partition_key,
                 row_key=row_key,
                 table_entity_properties=entity,