+++ sdk/textanalytics/azure-ai-textanalytics/tests/test_recognize_pii_entities_async.py
@@ -3,7 +3,7 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
-
+import os
 import pytest
 import platform
 import functools
@@ -17,6 +17,8 @@ from azure.ai.textanalytics.aio import TextAnalyticsClient
 from azure.ai.textanalytics import (
     TextDocumentInput,
     VERSION,
+    TextAnalyticsApiVersion,
+    PiiEntityDomainType,
 )
 
 # pre-apply the client_cls positional argument so it needn't be explicitly passed below
@@ -82,13 +84,6 @@ class TestRecognizePIIEntities(AsyncTextAnalyticsTest):
                 self.assertIsNotNone(entity.offset)
                 self.assertIsNotNone(entity.length)
                 self.assertIsNotNone(entity.confidence_score)
-
-    @GlobalTextAnalyticsAccountPreparer()
-    @TextAnalyticsClientPreparer()
-    async def test_length_with_emoji(self, client):
-        result = await client.recognize_pii_entities(["👩 SSN: 859-98-0987"])
-        self.assertEqual(result[0].entities[0].offset, 7)
-        self.assertEqual(result[0].entities[0].length, 11)
 
     @GlobalTextAnalyticsAccountPreparer()
     @TextAnalyticsClientPreparer()
@@ -570,3 +565,42 @@ class TestRecognizePIIEntities(AsyncTextAnalyticsTest):
             language="en",
             raw_response_hook=callback
         )
+
+    @GlobalTextAnalyticsAccountPreparer()
+    @TextAnalyticsClientPreparer(client_kwargs={"api_version": TextAnalyticsApiVersion.V3_0})
+    async def test_recognize_pii_entities_v3(self, client):
+        with pytest.raises(NotImplementedError) as excinfo:
+            await client.recognize_pii_entities(["this should fail"])
+
+        assert "'recognize_pii_entities' endpoint is only available for API version v3.1-preview.1 and up" in str(excinfo.value)
+
+    # currently only have this as playback since the dev endpoint is unreliable
+    @pytest.mark.playback_test_only
+    @GlobalTextAnalyticsAccountPreparer()
+    @TextAnalyticsClientPreparer(client_kwargs={
+        "api_version": TextAnalyticsApiVersion.V3_1_PREVIEW_2,
+        "text_analytics_account_key": os.environ.get('AZURE_TEXT_ANALYTICS_KEY'),
+        "text_analytics_account": "https://cognitiveusw2dev.azure-api.net/"
+    })
+    async def test_redacted_text(self, client):
+        result = await client.recognize_pii_entities(["My SSN is 859-98-0987."])
+        self.assertEqual("My SSN is ***********.", result[0].redacted_text)
+
+    @GlobalTextAnalyticsAccountPreparer()
+    @TextAnalyticsClientPreparer()
+    async def test_redacted_text_v3_1_preview_1(self, client):
+        result = await client.recognize_pii_entities(["My SSN is 859-98-0987."])
+        self.assertIsNone(result[0].redacted_text)
+        
+    @GlobalTextAnalyticsAccountPreparer()
+    @TextAnalyticsClientPreparer()
+    async def test_phi_domain_filter(self, client):
+        # without the domain filter, this should return two entities: Microsoft as an org,
+        # and the phone number. With the domain filter, it should only return one.
+        result = await client.recognize_pii_entities(
+            ["I work at Microsoft and my phone number is 333-333-3333"],
+            domain_filter=PiiEntityDomainType.PROTECTED_HEALTH_INFORMATION
+        )
+        self.assertEqual(len(result[0].entities), 1)
+        self.assertEqual(result[0].entities[0].text, '333-333-3333')
+        self.assertEqual(result[0].entities[0].category, 'Phone Number')